refactor: extract interface IVoxelmapCollider
