+++ src/lib/physics/i-voxelmap-collider.ts
@@ -1,0 +1,13 @@
+import type * as THREE from '../libs/three-usage';
+
+enum EVoxelStatus {
+    EMPTY,
+    FULL,
+    NOT_LOADED,
+}
+
+interface IVoxelmapCollider {
+    getVoxel(worldVoxelCoords: THREE.Vector3Like): EVoxelStatus;
+}
+
+export { EVoxelStatus, type IVoxelmapCollider };
